+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Next Release
 
+* **Admin Fix**: Fix error in Importing under Postgres
 * **Admin Feature** Ability to add entity to LIS Added
 * Added code endpoint for LIS organisations for web app use
 * Schema updated to account for these changes